OSC52 copying in Kitty
This commit allows you to actually copy content from nvim over ssh in
kitty. For some strange reason it doesn't work out of the box with nvim.
The relevant thread can be found here https://github.com/LazyVim/LazyVim/discussions/2715
and it should be noted that I didn't need to edit the kitty config but
that should be tried if this change doesn't work.